requirements.txt
hvac==0.6.1
pyyaml>=5
python-ldap==3.2.0
coloredlogs==10.0
tox==3.13.2
tox-docker==1.4.1
pytest==4.5.0
Jinja2==2.10.1
